loader/requirements.txt
faker==24.9.0
psycopg2-binary==2.9.9
